Move the method to the DC_Table class
