+++ calendar-bundle/src/Resources/contao/dca/tl_calendar_events.php
@@ -1146,7 +1146,7 @@ class tl_calendar_events extends Contao\Backend
 		// The onsubmit_callback has triggered scheduleUpdate(), so run generateFeed() now
 		$this->generateFeed();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
