+++ comments-bundle/src/Resources/contao/dca/tl_comments.php
@@ -745,7 +745,7 @@ class tl_comments extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
