+++ core-bundle/src/Resources/contao/dca/tl_article.php
@@ -1006,7 +1006,7 @@ class tl_article extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
