+++ core-bundle/src/Resources/contao/dca/tl_content.php
@@ -2097,7 +2097,7 @@ class tl_content extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
