+++ core-bundle/src/Resources/contao/dca/tl_form_field.php
@@ -908,7 +908,7 @@ class tl_form_field extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
