+++ core-bundle/src/Resources/contao/dca/tl_image_size_item.php
@@ -394,7 +394,7 @@ class tl_image_size_item extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
