+++ core-bundle/src/Resources/contao/dca/tl_member_group.php
@@ -341,7 +341,7 @@ class tl_member_group extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
