+++ core-bundle/src/Resources/contao/dca/tl_style.php
@@ -801,7 +801,7 @@ class tl_style extends Contao\Backend
 		// The onsubmit_callback has triggered scheduleUpdate(), so run updateStyleSheet() now (see #1052)
 		$this->updateStyleSheet();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
