+++ core-bundle/src/Resources/contao/dca/tl_user.php
@@ -1090,7 +1090,7 @@ class tl_user extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
