+++ core-bundle/src/Resources/contao/dca/tl_user_group.php
@@ -585,7 +585,7 @@ class tl_user_group extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
