+++ core-bundle/src/Resources/contao/drivers/DC_Table.php
@@ -14,6 +14,7 @@ use Contao\CoreBundle\Exception\AccessDeniedException;
 use Contao\CoreBundle\Exception\InternalServerErrorException;
 use Contao\CoreBundle\Exception\ResponseException;
 use Contao\CoreBundle\Picker\PickerInterface;
+use FOS\HttpCacheBundle\CacheManager;
 use Patchwork\Utf8;
 use Symfony\Component\HttpFoundation\Session\Attribute\AttributeBagInterface;
 use Symfony\Component\HttpFoundation\Session\Session;
@@ -6179,6 +6180,46 @@ class DC_Table extends DataContainer implements \listable, \editable
 		}
 
 		return $attributes;
+	}
+
+	/**
+	 * Invalidate the cache tags associated with a given DC
+	 *
+	 * Call this whenever an entry is modified (added, updated, deleted).
+	 */
+	public function invalidateCacheTags()
+	{
+		if (!System::getContainer()->has('fos_http_cache.cache_manager'))
+		{
+			return;
+		}
+
+		$ns = 'contao.db.';
+		$tags = array($ns . $this->table, $ns . $this->table . '.' . $this->id);
+
+		// Trigger the oninvalidate_cache_tags_callback
+		if (\is_array($GLOBALS['TL_DCA'][$this->table]['config']['oninvalidate_cache_tags_callback']))
+		{
+			foreach ($GLOBALS['TL_DCA'][$this->table]['config']['oninvalidate_cache_tags_callback'] as $callback)
+			{
+				if (\is_array($callback))
+				{
+					$this->import($callback[0]);
+					$tags = $this->{$callback[0]}->{$callback[1]}($this, $tags);
+				}
+				elseif (\is_callable($callback))
+				{
+					$tags = $callback($this, $tags);
+				}
+			}
+		}
+
+		// Make sure tags are unique and empty ones are removed
+		$tags = array_filter(array_unique($tags));
+
+		/** @var CacheManager $cacheManager */
+		$cacheManager = System::getContainer()->get('fos_http_cache.cache_manager');
+		$cacheManager->invalidateTags($tags);
 	}
 }
