+++ faq-bundle/src/Resources/contao/dca/tl_faq.php
@@ -672,7 +672,7 @@ class tl_faq extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
