+++ newsletter-bundle/src/Resources/contao/dca/tl_newsletter_recipients.php
@@ -502,7 +502,7 @@ class tl_newsletter_recipients extends Contao\Backend
 
 		$objVersions->create();
 
-		if ($dc)
+		if ($dc instanceof Contao\DC_Table)
 		{
 			$dc->invalidateCacheTags();
 		}
